re-run make idl after Jelmer's "poperly cast array lengths" pidl change.
(f321240fa91fa19c1131f119c42f64897d220682)

Michael
(This used to be commit 962c2670b10addf81a242d8cb381cc010e036b8e)

+++ source3/librpc/gen_ndr/ndr_dfs.c
@@ -503,7 +503,7 @@ _PUBLIC_ void ndr_print_dfs_Info3(struct ndr_print *ndr, const char *name, const
 	ndr_print_ptr(ndr, "stores", r->stores);
 	ndr->depth++;
 	if (r->stores) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "stores", r->num_stores);
+		ndr->print(ndr, "%s: ARRAY(%d)", "stores", (int)r->num_stores);
 		ndr->depth++;
 		for (cntr_stores_1=0;cntr_stores_1<r->num_stores;cntr_stores_1++) {
 			char *idx_1=NULL;
@@ -666,7 +666,7 @@ _PUBLIC_ void ndr_print_dfs_Info4(struct ndr_print *ndr, const char *name, const
 	ndr_print_ptr(ndr, "stores", r->stores);
 	ndr->depth++;
 	if (r->stores) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "stores", r->num_stores);
+		ndr->print(ndr, "%s: ARRAY(%d)", "stores", (int)r->num_stores);
 		ndr->depth++;
 		for (cntr_stores_1=0;cntr_stores_1<r->num_stores;cntr_stores_1++) {
 			char *idx_1=NULL;
@@ -1074,7 +1074,7 @@ _PUBLIC_ void ndr_print_dfs_Info6(struct ndr_print *ndr, const char *name, const
 	ndr_print_ptr(ndr, "stores", r->stores);
 	ndr->depth++;
 	if (r->stores) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "stores", r->num_stores);
+		ndr->print(ndr, "%s: ARRAY(%d)", "stores", (int)r->num_stores);
 		ndr->depth++;
 		for (cntr_stores_1=0;cntr_stores_1<r->num_stores;cntr_stores_1++) {
 			char *idx_1=NULL;
@@ -2265,7 +2265,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray1(struct ndr_print *ndr, const char *name, 
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2351,7 +2351,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray2(struct ndr_print *ndr, const char *name, 
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2437,7 +2437,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray3(struct ndr_print *ndr, const char *name, 
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2523,7 +2523,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray4(struct ndr_print *ndr, const char *name, 
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2609,7 +2609,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray5(struct ndr_print *ndr, const char *name, 
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2695,7 +2695,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray6(struct ndr_print *ndr, const char *name, 
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2781,7 +2781,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray200(struct ndr_print *ndr, const char *name
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
@@ -2867,7 +2867,7 @@ _PUBLIC_ void ndr_print_dfs_EnumArray300(struct ndr_print *ndr, const char *name
 	ndr_print_ptr(ndr, "s", r->s);
 	ndr->depth++;
 	if (r->s) {
-		ndr->print(ndr, "%s: ARRAY(%d)", "s", r->count);
+		ndr->print(ndr, "%s: ARRAY(%d)", "s", (int)r->count);
 		ndr->depth++;
 		for (cntr_s_1=0;cntr_s_1<r->count;cntr_s_1++) {
 			char *idx_1=NULL;
